mm, sl[au]b: Introduce lockless cache
This is v3 of lockless cache on slab. Note that this is not
final form of v3. There are some issues to solve. Also note
that it's not implemented in SLAB yet.

v2 is here:
        - https://lkml.org/lkml/2021/9/20/600

Change from v2:
        - Add SLAB_LOCKLESS_CACHE in SLAB_CORE_FLAGS. without this,
                slab fails to create a cache.
        - Allocate and free with kmem_cache_{alloc,free}_bulk in slowpath.
        - Use {get,put}_cpu API to avoid preemption during allocation.
	- Fix memory leak on shutdown_cache
	- Process __GFP_ZERO as ignoring that can be problematic.

code snippet below is ugly but it has a reason.  just calling
kmem_cache_alloc_bulk with GFP_ATOMIC can crash when gfpflags
is more restrictive than GFP_ATOMIC.

+               gfpflags &= GFP_ATOMIC;
+               if (!gfpflags)
+                       gfpflags = GFP_ATOMIC;
+
+               cache->size += kmem_cache_alloc_bulk(s, gfpflags,
+                               KMEM_LOCKLESS_CACHE_BATCHCOUNT,
+                               cache->queue);

Many thanks to Matthew, Vlastimil, Jens and John who gave feedback.

Signed-off-by: Hyeonggon Yoo <42.hyeyoo@gmail.com>