x86/mm/cpa: do not allow working with user mappings
Currently CPA is not used for user mappings (only pgd of init_mm
or and efi_mm is used). For simplicity, do not allow working with
user mappings.

Let CPA fail when one of following conditions are true:
	- user address is passed
	- old pte is with _PAGE_PRESENT|_PAGE_USER
	- new pte result in _PAGE_PRESENT|_PAGE_USER

Note that efi_mm uses 1:1 mapping so its address should not be
considered as user address.

Signed-off-by: Hyeonggon Yoo <42.hyeyoo@gmail.com>